requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
defusedxml==0.7.1
distro==1.9.0
Django==5.2.1
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.0
django-taggit==6.1.0
django-tasks==0.6.1
django-treebeard==4.7.1
djangorestframework==3.16.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
laces==0.1.2
openai==1.82.1
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
pillow_heif==0.21.0
pydantic==2.11.5
pydantic_core==2.33.2
python-decouple==3.8
python-dotenv==1.1.0
python-telegram-bot==22.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wagtail==6.4.1
whitenoise==6.9.0
Willow==1.10.0
